+++ src/parsley/field.js
@@ -275,11 +275,10 @@ ParsleyField.prototype = {
 
     // Small special case here for HTML5 number: integer validator if step attribute is undefined or an integer value, number otherwise
     if ('number' === type) {
-      if (('undefined' === typeof this.$element.attr('step')) || (0 === parseFloat(this.$element.attr('step')) % 1)) {
-        return this.addConstraint('type', 'integer', undefined, true);
-      } else {
-        return this.addConstraint('type', 'number', undefined, true);
-      }
+      return this.addConstraint('type', ['number', {
+        step: this.$element.attr('step'),
+        base: this.$element.attr('min') || this.$element.attr('value')
+      }], undefined, true);
     // Regular other HTML5 supported types
     } else if (/^(email|url|range)$/i.test(type)) {
       return this.addConstraint('type', type, undefined, true);