+++ src/parsley/validator_registry.js
@@ -15,7 +15,8 @@ var ParsleyValidatorRegistry = function (validators, catalog) {
 var typeRegexes =  {
   email: /^((([a-z]|\d|[!#\$%&'\*\+\-\/=\?\^_`{\|}~]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])+(\.([a-z]|\d|[!#\$%&'\*\+\-\/=\?\^_`{\|}~]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])+)*)|((\x22)((((\x20|\x09)*(\x0d\x0a))?(\x20|\x09)+)?(([\x01-\x08\x0b\x0c\x0e-\x1f\x7f]|\x21|[\x23-\x5b]|[\x5d-\x7e]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])|(\\([\x01-\x09\x0b\x0c\x0d-\x7f]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF]))))*(((\x20|\x09)*(\x0d\x0a))?(\x20|\x09)+)?(\x22)))@((([a-z]|\d|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])|(([a-z]|\d|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])([a-z]|\d|-|\.|_|~|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])*([a-z]|\d|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])))\.)+(([a-z]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])|(([a-z]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])([a-z]|\d|-|\.|_|~|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])*([a-z]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])))$/i,
 
-  number: /^-?(?:(?:\d+|\d{1,3}(?:,\d{3})+)+(?:\.\d+)?|(?:\.\d+)+)$/,
+  // Follow https://www.w3.org/TR/html5/infrastructure.html#floating-point-numbers
+  number: /^-?(\d*\.)?\d+(e[-+]?\d+)?$/i,
 
   integer: /^-?\d+$/,
 
@@ -59,6 +60,18 @@ var typeRegexes =  {
     )
 };
 typeRegexes.range = typeRegexes.number;
+
+// See http://stackoverflow.com/a/10454560/8279
+var decimalPlaces = num => {
+  var match = ('' + num).match(/(?:\.(\d+))?(?:[eE]([+-]?\d+))?$/);
+  if (!match) { return 0; }
+  return Math.max(
+       0,
+       // Number of digits right of decimal point.
+       (match[1] ? match[1].length : 0) -
+       // Adjust for scientific notation.
+       (match[2] ? +match[2] : 0));
+};
 
 ParsleyValidatorRegistry.prototype = {
   init: function (validators, catalog) {
@@ -220,11 +233,28 @@ ParsleyValidatorRegistry.prototype = {
       priority: 512
     },
     type: {
-      validateString: function(value, type) {
+      validateString: function(value, type, {step = '1', base = 0} = {}) {
         var regex = typeRegexes[type];
-        if (!regex)
+        if (!regex) {
           throw new Error('validator type `' + type + '` is not supported');
-        return regex.test(value);
+        }
+        if (!regex.test(value))
+          return false;
+        if ('number' === type) {
+          if (!/^any$/i.test(step || '')) {
+            var nb = Number(value);
+            // Be careful of rounding errors by using integers.
+            var mul = Math.pow(10, Math.max(decimalPlaces(step), decimalPlaces(base)));
+            if ((nb * mul - base * mul) % (step * mul) != 0)
+              return false;
+          }
+        }
+        return true;
+      },
+      requirementType: {
+        '': 'string',
+        step: 'string',
+        base: 'number'
       },
       priority: 256
     },