+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## master
 - Drop support for undocumented option `eventValidate`.
+- type="number" now follows HTML5 spec. In particular, commas are no longer accepted. (#1037)
 
 ## 2.2.0-rc3